requirements/ci.txt
-r prod.txt
coverage==3.7.1
coveralls==0.5
behave==1.2.5
inflection==0.3.1
django-extensions==1.5.5
django-autofixture==0.10.1
jsonschema==2.5.1
requests==2.7.0
sure==1.2.13
